[notebook][hailtop][gear] move logging configuration out of gear
`gear` should *only* be used in services, *never* in developer or user facing
tools. This change fixes `scale-test.py` to not use `gear`.
