+++ batch/batch/front_end/front_end.py
@@ -20,10 +20,10 @@ from hailtop.utils import (time_msecs, time_msecs_str, humanize_timedelta_msecs,
 from hailtop.batch_client.parse import parse_cpu_in_mcpu, parse_memory_in_bytes
 from hailtop.config import get_deploy_config
 from hailtop.tls import get_in_cluster_server_ssl_context, in_cluster_ssl_client_session
+from hailtop.hail_logging import AccessLogger
 from gear import (Database, setup_aiohttp_session,
                   rest_authenticated_users_only, web_authenticated_users_only,
-                  web_authenticated_developers_only, check_csrf_token, transaction,
-                  AccessLogger)
+                  web_authenticated_developers_only, check_csrf_token, transaction)
 from web_common import (setup_aiohttp_jinja2, setup_common_static_routes,
                         render_template, set_message)
