+++ hail/python/hailtop/hail_logging.py
