+++ notebook/notebook/notebook.py
@@ -13,11 +13,12 @@ import kubernetes_asyncio as kube
 
 from hailtop.config import get_deploy_config
 from hailtop.tls import get_in_cluster_server_ssl_context
-from gear import setup_aiohttp_session, create_database_pool, \
-    web_authenticated_users_only, web_maybe_authenticated_user, web_authenticated_developers_only, \
-    check_csrf_token, AccessLogger
-from web_common import sass_compile, setup_aiohttp_jinja2, setup_common_static_routes, \
-    set_message, render_template
+from hailtop.hail_logging import AccessLogger
+from gear import (setup_aiohttp_session, create_database_pool,
+                  web_authenticated_users_only, web_maybe_authenticated_user,
+                  web_authenticated_developers_only, check_csrf_token)
+from web_common import (sass_compile, setup_aiohttp_jinja2,
+                        setup_common_static_routes, set_message, render_template)
 
 log = logging.getLogger('notebook')
