+++ query/query/query.py
@@ -9,7 +9,8 @@ from py4j.java_gateway import JavaGateway, GatewayParameters, launch_gateway
 from hailtop.utils import blocking_to_async, retry_transient_errors
 from hailtop.config import get_deploy_config
 from hailtop.tls import get_in_cluster_server_ssl_context
-from gear import setup_aiohttp_session, rest_authenticated_users_only, AccessLogger
+from hailtop.hail_logging import AccessLogger
+from gear import setup_aiohttp_session, rest_authenticated_users_only
 
 uvloop.install()
