+++ router-resolver/router_resolver/router_resolver.py
@@ -7,7 +7,8 @@ from kubernetes_asyncio import client, config
 import logging
 from hailtop.auth import async_get_userinfo
 from hailtop.tls import get_in_cluster_server_ssl_context
-from gear import configure_logging, setup_aiohttp_session
+from hailtop.hail_logging import configure_logging
+from gear import setup_aiohttp_session
 
 uvloop.install()
